Breaking change update update
Ignore the last commit, apparently it was a PEBKAC error. However, for clarification on my last post, I recommend copying serverdict and serverdict_backup, renaming copies to oldserverdict and oldserverdict_backup. Then, run the script and rename serverdict_converted to serverdict. 

If you are having issues with the converter, just reconfigure. This is only an attempt to try to avoid reconfiguring, it isn't necessary.

This is only necessary for those currently running the script, newcomers should be fine.