!starttime added
If anybody is still having issues with the converter, it has to be ran with the bot NOT running. If you are getting errors, it is best to close the bot, delete data/serverdict and data/serverdict_backup, start the bot up again and reconfigure. I apologize, but the change is worth it going forward.