Added !starting <team>, Auto-delete !invite messages, perm/emoji fixes
Added the ability to start raids by team. This is a conversion of the guild_dict structure and !convert has been added to convert to the new structure. This is a temporary function and will need to be ran as soon as you load up this update, and only once. All new raids should work fine, this is to fix currently active raids, such as EX raids. If you are worried about it, wait until off-hours after EX raids have finished. I will remove !convert soon