Begin moving to human readable guild_dict.json. SEE NOTES  BrenenP committed
the old pickled serverdict file is not human readable and there are disadvantages to pickle. I am beginning to investigate moving to JSON instead. This commit makes a backup file that is unused for now in JSON format. This file is a human-readable version of guild_dict. In the coming weeks I may start to use this as primary storage